% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{compress_fastq}
\alias{compress_fastq}
\title{Compress FASTQ files using pigz}
\usage{
compress_fastq(fastq_files, threads = 8, software_paths)
}
\arguments{
\item{fastq_files}{Character vector of FASTQ file paths (.fastq).}

\item{threads}{Number of threads.}

\item{software_paths}{Named list of software paths, especially \code{pigz}.}

\item{remove_original}{Remove original FASTQ after compression?}
}
\value{
Character vector of compressed file paths (.fastq.gz).

Character vector of successfully compressed file paths (.fastq.gz).
}
\description{
Compress FASTQ files using pigz
}
\keyword{Compress}
\keyword{FASTQ}
\keyword{files}
\keyword{internal}
\keyword{pigz}
\keyword{using}
