% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{convert_sra_to_fastq}
\alias{convert_sra_to_fastq}
\title{Convert SRA to FASTQ using fasterq-dump}
\usage{
convert_sra_to_fastq(sra_file, output_dir = ".", threads = 8, software_paths)
}
\arguments{
\item{sra_file}{Path to the SRA file.}

\item{output_dir}{Directory to write FASTQ files.}

\item{threads}{Number of threads to use.}

\item{software_paths}{Paths to tools (esp. fasterq-dump).}
}
\value{
Character vector of generated FASTQ file paths.
}
\description{
Convert SRA to FASTQ using fasterq-dump
}
\keyword{internal}
