% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ena_sra_run}
\alias{download_ena_sra_run}
\title{Download a Single ENA/SRA Run}
\usage{
download_ena_sra_run(
  run_id,
  run_info,
  download_fastq_gz = TRUE,
  use_aspera = FALSE,
  database_preference = "ena",
  parallel_n = 0,
  output_dir = ".",
  software_paths
)
}
\arguments{
\item{run_id}{SRR/ERR/DRR identifier.}

\item{run_info}{A single row tibble/data.frame from the metadata file for this run.}

\item{download_fastq_gz}{Should FASTQ.gz be preferred?}

\item{use_aspera}{Should Aspera be used if available?}

\item{database_preference}{"ena" or "sra".}

\item{parallel_n}{Parallel connections for axel.}

\item{output_dir}{Destination directory.}

\item{software_paths}{Paths to tools.}
}
\value{
A list containing paths to downloaded files (or the SRA file path). Keys: \code{files}, \code{type} ("fastq_gz", "sra").
}
\description{
Handles logic for downloading FASTQ or SRA, using FTP/HTTPS/Aspera.
}
\keyword{internal}
