% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{execute_aspera}
\alias{execute_aspera}
\title{Download Helper using Aspera}
\usage{
execute_aspera(
  remote_path,
  db = "ENA",
  software_paths,
  aspera_key_path,
  output_dir = "."
)
}
\arguments{
\item{remote_path}{The Aspera remote path (e.g., era-fasp@...).}

\item{db}{Source database ("ENA" or "GSA").}

\item{software_paths}{List containing path to 'ascp'.}

\item{aspera_key_path}{Path to the Aspera key file.}

\item{output_dir}{Local directory to download into.}
}
\description{
Download Helper using Aspera
}
\keyword{internal}
