% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{execute_download}
\alias{execute_download}
\title{Download Helper using wget or axel}
\usage{
execute_download(url, dest_file, parallel_n = 0, software_paths)
}
\arguments{
\item{url}{The URL to download.}

\item{dest_file}{The destination file path.}

\item{parallel_n}{Number of parallel connections for axel (0 means use wget).}

\item{software_paths}{List containing paths to 'wget' and 'axel'.}
}
\description{
Download Helper using wget or axel
}
\keyword{internal}
