% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_aspera_key_path}
\alias{get_aspera_key_path}
\title{Get Aspera Key Path}
\usage{
get_aspera_key_path(db = "ENA", software_paths)
}
\arguments{
\item{db}{"ENA" or "GSA".}

\item{software_paths}{List containing path to 'ascp'.}
}
\value{
Path to the key file or NULL.
}
\description{
Finds the appropriate Aspera key file based on standard locations or options.
}
\keyword{internal}
