% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Metadata Wrapper}
\usage{
get_metadata(accession, output_dir = ".")
}
\arguments{
\item{accession}{Accession ID.}

\item{output_dir}{Output directory.}
}
\value{
Path to the metadata file.
}
\description{
Determines the source (GSA or ENA/SRA) and calls the appropriate function.
}
\keyword{internal}
