% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_software_path}
\alias{get_software_path}
\title{Get Path to External Software}
\usage{
get_software_path(software_name, package_name = software_name, required = TRUE)
}
\arguments{
\item{software_name}{The name of the executable (e.g., "wget").}

\item{package_name}{Optional: Name of the package/tool for error messages (e.g., "wget", "sra-tools").}

\item{required}{Logical: If TRUE, throws an error if not found. If FALSE, returns NULL.}
}
\value{
The path to the executable or NULL/Error.
}
\description{
Checks PATH environment variable and R options for software executable.
}
\keyword{internal}
