% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseq_download.R
\name{init_software_paths}
\alias{init_software_paths}
\title{Check and initialize required software paths}
\usage{
init_software_paths(gzip, fastq, merge, parallel, aspera)
}
\description{
Check and initialize required software paths
}
\keyword{internal}
