% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_already_successful}
\alias{is_already_successful}
\title{Check if Run is Already Successfully Downloaded}
\usage{
is_already_successful(run_id, log_dir = ".")
}
\arguments{
\item{run_id}{The Run accession.}

\item{log_dir}{Directory where logs are stored.}
}
\value{
Logical TRUE if found in success.log.
}
\description{
Check if Run is Already Successfully Downloaded
}
\keyword{internal}
