% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseq_download.R
\name{iseq_download}
\alias{iseq_download}
\title{Main function}
\usage{
iseq_download(
  input,
  output_dir = ".",
  metadata_only = FALSE,
  gzip = FALSE,
  fastq = FALSE,
  merge = c("none", "ex", "sa", "st"),
  threads = 8,
  database = c("ena", "sra"),
  parallel = 0,
  aspera = FALSE,
  retries = 2,
  remove_sra = TRUE
)
}
\description{
Main function
}
