% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_status}
\alias{log_status}
\title{Log Download Status}
\usage{
log_status(run_id, status = "success", log_dir = ".")
}
\arguments{
\item{run_id}{The Run accession (e.g., SRR12345).}

\item{status}{"success" or "fail".}

\item{log_dir}{Directory where logs are stored.}
}
\description{
Appends run ID to success or fail log file.
}
\keyword{internal}
