% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseq_download.R
\name{post_process_files}
\alias{post_process_files}
\title{Post-process downloaded files}
\usage{
post_process_files(run_id, downloaded_info, params, software_paths)
}
\description{
Post-process downloaded files
}
\keyword{internal}
