% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseq_download.R
\name{process_metadata}
\alias{process_metadata}
\title{Process metadata for a single accession}
\usage{
process_metadata(accession, output_dir)
}
\description{
Process metadata for a single accession
}
\keyword{internal}
