% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_command}
\alias{run_command}
\title{Run External Command}
\usage{
run_command(
  command,
  args = character(),
  error_on_status = TRUE,
  spinner = FALSE,
  ...
)
}
\arguments{
\item{command}{Path to the command.}

\item{args}{Character vector of arguments.}

\item{error_on_status}{Should an error be thrown on non-zero exit status?}

\item{spinner}{Show a spinner?}

\item{...}{Additional arguments passed to processx::run.}
}
\value{
Result from processx::run.
}
\description{
Wrapper around processx::run for consistency and error handling.
}
\keyword{internal}
