% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{validate_and_query_ena}
\alias{validate_and_query_ena}
\title{Validate Accession and Generate ENA Query}
\usage{
validate_and_query_ena(accession)
}
\arguments{
\item{accession}{Input accession (Project, Study, Sample, Exp, Run, GSE, GSM).}
}
\value{
ENA query string or throws error.
}
\description{
Validate Accession and Generate ENA Query
}
\keyword{internal}
